Reorganize notification helper code.
