+++ mobile/src/main/java/org/openhab/habdroid/ui/MainActivity.kt
@@ -70,7 +70,6 @@ import org.openhab.habdroid.R
 import org.openhab.habdroid.background.BackgroundTasksManager
 import org.openhab.habdroid.background.NotificationUpdateObserver
 import org.openhab.habdroid.core.CloudMessagingHelper
-import org.openhab.habdroid.core.MessageListenerService.Companion.checkForMessages
 import org.openhab.habdroid.core.UpdateBroadcastReceiver
 import org.openhab.habdroid.core.VoiceService
 import org.openhab.habdroid.core.connection.CloudConnection
@@ -332,7 +331,7 @@ class MainActivity : AbstractBaseActivity(), ConnectionFactory.UpdateListener {
         return when (item.itemId) {
             R.id.mainmenu_voice_recognition -> {
                 launch {
-                    checkForMessages(this@MainActivity)
+                    CloudMessagingHelper.pollForNotifications(this@MainActivity)
                 }
                 launchVoiceRecognition()
                 true
